+++ srcs/ray.h
@@ -6,7 +6,7 @@
 /*   By: lchapren <lchapren@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/08 12:57:51 by lchapren          #+#    #+#             */
-/*   Updated: 2020/06/17 12:58:39 by lchapren         ###   ########.fr       */
+/*   Updated: 2020/07/01 11:47:22 by lchapren         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,8 @@ typedef struct	s_ray
 
 	int		step_x;
 	int		step_y;
-	int		map_x;
-	int		map_y;
+	float		map_x;
+	float		map_y;
 	int		wall_hit;
 	int		hit_side;
 	int		column;
